Now the program computes the root of any function.
The main bug was that a pointer was assigned into another one when the
intent was to copy the memory it was pointing to. Also, it didn't make
any sense to use the norm of the initial point as the error.

There's still no proper error messaging and user options.